requirements.txt
torch==1.10.0
numpy==1.21.2
Pillow==8.4.0